Numpad Layer Adjustments
Numpad becomes

   7 8 9 ,
 - 4 5 6 0
   1 2 3 .
   [ e ]

The left hand then gets ctrl and alt at positions easier to press
together with the left thumb.

An additional Numpad layer is introduced as 6th (good grief) layer,
positioned lefthand

 X789
 -456+
 /123*
 Y  D0

where X returns to layer 0 and Y is enter. D is the decimal point, which
will be interpreted as ','