Add Repeat key/Mousing, Remove Tap Dance/Sendstring
Number/Nav Layer becomes a bit full now:

Right Hand:

        x 7 8 9 :
        [ 4 5 6 0
      P ] 1 2 3 .
      N S ( , ) e
  P: print
  N: toggles this Layer
  S: space

  x and e are included, because thes enables one to write 4e-4 and 2*x

Left Hand:

  HOME PGUP UP   PGDW END
  BKSP LE   DW   RGT  DEL
  ALT  GUI  SFT  CTL  $    %
       *    /    -    +    X

  X: to third layer

  $ is included beause you now can write $1/$5, not really beause of currency

mousing is added back into the third layer, because I'm now using a trackpad,
and using the mouse keys of the trackpad is awkward. For easy acces, a toggle
key is added to the base layer.

Having Callum mods (Modifiers on the home row in an upper layer, or in my case
below the home row) makes it awkward to retype key combos like <c-+> or <a-,>.
or something like <g-c-up>. A repeat key, together with an anti-repeat key
should make it easier to type this. I'll also try to incorporate it in my typing
of words with double letters.