add back number/nav layer
I was getting annoyed with putting in numbers via the symbol layer,
so I'm adding back in a number layer. I was also getting annoyed with
the combos for CTL and ALT in the base layer and found that I wasn't using
REP/AREP enough to justify this.

So I moved (A)REP to the outside and moved the modifiers.

The new keypad layer has a shift key to make the numbers be F keys and a
lock key to make it keypad keys (needed for software that behaves differently if a keypad
is used) these spring back to number row behavior if the layer is left.