Base layer nav & num-cluster
move num-cluster activation one to the left for OÖ to EÄ.
this is where KC_RGHT was, so move this and KC_LEFT one up (TV -> DT, EÄ -> AE)
this is where KC_PGUP/KC_PGDN were, so move these to the right (XO/OÖ)

    le up ri  pu
       dn N   pd

Positions of KC_HOME/KC_END stay where they were (now le+up / up+ri).

Also add the EÄ positioned lNUM activation to the symbol layer.
This should make it easier to type equations and timestamps.

On the number layer itself, change the activation of function keys from the
ring finger to the pinky.